Rename components to include their scope
